UI : factor thermal/renewable area summary classes
Addition of a common area summary class that contains common data and behaviors of thermal/renewable area summary classes
