+++ src/ui/simulator/toolbox/components/datagrid/renderer/area/renewable.areasummary.cpp
@@ -26,8 +26,6 @@
 */
 
 #include "renewable.areasummary.h"
-#include "../../../refresh.h"
-#include "../../../../../application/study.h"
 
 using namespace Yuni;
 
@@ -40,18 +38,11 @@ namespace Datagrid
 namespace Renderer
 {
 RenewableClusterSummarySingleArea::RenewableClusterSummarySingleArea(
-  wxWindow* control,
-  Toolbox::InputSelector::Area* notifier) :
- pArea(nullptr), pControl(control), pAreaNotifier(notifier)
-{
-    if (notifier)
-        notifier->onAreaChanged.connect(this, &RenewableClusterSummarySingleArea::onAreaChanged);
-}
+    wxWindow* control,
+    Toolbox::InputSelector::Area* notifier) : CommonClusterSummarySingleArea(control, notifier)
+{}
 
-RenewableClusterSummarySingleArea::~RenewableClusterSummarySingleArea()
-{
-    destroyBoundEvents();
-}
+RenewableClusterSummarySingleArea::~RenewableClusterSummarySingleArea() {}
 
 wxString RenewableClusterSummarySingleArea::rowCaption(int rowIndx) const
 {
@@ -114,102 +105,6 @@ double RenewableClusterSummarySingleArea::cellNumericValue(int x, int y) const
     return 0.;
 }
 
-void RenewableClusterSummarySingleArea::onAreaChanged(Antares::Data::Area* area)
-{
-    if (pArea != area)
-    {
-        pArea = area;
-        RefreshAllControls(pControl);
-    }
-}
-
-IRenderer::CellStyle RenewableClusterSummarySingleArea::cellStyle(int col, int row) const
-{
-    if (col > 0 and Math::Zero(cellNumericValue(col, row)))
-        return IRenderer::cellStyleDefaultDisabled;
-    else
-    {
-        if (col == 1 || col == 2)
-            return IRenderer::cellStyleConstraintWeight;
-        else
-            return IRenderer::cellStyleDefault;
-    }
-}
-
-struct NoCheck
-{
-    template<class T>
-    static bool Validate(const T&)
-    {
-        return true;
-    }
-};
-
-struct CheckUnitCount
-{
-    static bool Validate(uint& f)
-    {
-        if (f > 100)
-            f = 100;
-        return true;
-    }
-};
-
-template<class CheckT>
-static bool UpdateUnsignedLong(uint& value, const String& str)
-{
-    uint l;
-    if (str.to(l))
-    {
-        if (value != l and CheckT::Validate(l))
-        {
-            value = l;
-            MarkTheStudyAsModified();
-            OnInspectorRefresh(nullptr);
-            return true;
-        }
-    }
-    return false;
-}
-
-template<class CheckT>
-static bool UpdateDouble(double& value, const String& str)
-{
-    double d;
-    if (str.to(d))
-    {
-        if (not Math::Equals<double>(value, d))
-        {
-            if (CheckT::Validate(d))
-            {
-                value = d;
-                OnInspectorRefresh(nullptr);
-                return true;
-            }
-        }
-    }
-    return false;
-}
-
-template<class CheckT>
-static bool UpdateBool(bool& value, const String& str)
-{
-    bool d;
-    if (str.to(d))
-    {
-        if (value != d)
-        {
-            if (CheckT::Validate(d))
-            {
-                value = d;
-                OnInspectorRefresh(nullptr);
-                return true;
-            }
-        }
-    }
-    return false;
-}
-
 bool RenewableClusterSummarySingleArea::cellValue(int x, int y, const String& v)
 {
     auto* cluster = (pArea and (uint) y < pArea->renewable.list.size())
@@ -238,18 +133,6 @@ bool RenewableClusterSummarySingleArea::cellValue(int x, int y, const String& v)
         }
     }
     return false;
-}
-
-void RenewableClusterSummarySingleArea::onStudyClosed()
-{
-    pArea = nullptr;
-    IRenderer::onStudyClosed();
-}
-
-void RenewableClusterSummarySingleArea::onStudyAreaDelete(Antares::Data::Area* area)
-{
-    if (pArea == area)
-        onAreaChanged(nullptr);
 }
 
 } // namespace Renderer
