+++ src/ui/simulator/toolbox/components/datagrid/renderer/area/renewable.areasummary.h
@@ -27,10 +27,7 @@
 #ifndef __ANTARES_TOOLBOX_COMPONENT_DATAGRID_RENDERER_AREA_RENEWABLE_SUMMARY_SINGLE_AREA_H__
 #define __ANTARES_TOOLBOX_COMPONENT_DATAGRID_RENDERER_AREA_RENEWABLE_SUMMARY_SINGLE_AREA_H__
 
-// #include <antares/wx-wrapper.h>
-#include "../../renderer.h"
-#include "../../../../input/area.h"
-// #include <antares/study/parts/thermal/cluster.h>
+#include "common.areasummary.h"
 
 namespace Antares
 {
@@ -40,7 +37,7 @@ namespace Datagrid
 {
 namespace Renderer
 {
-class RenewableClusterSummarySingleArea : public IRenderer
+class RenewableClusterSummarySingleArea : public CommonClusterSummarySingleArea
 {
 public:
     RenewableClusterSummarySingleArea(wxWindow* control, Toolbox::InputSelector::Area* notifier);
@@ -56,18 +53,20 @@ public:
         return (pArea) ? pArea->renewable.list.size() : 0;
     }
 
-    virtual wxString columnCaption(int colIndx) const;
+    virtual wxString columnCaption(int colIndx) const override;
 
-    virtual wxString rowCaption(int rowIndx) const;
+    virtual wxString rowCaption(int rowIndx) const override;
 
-    virtual wxString cellValue(int x, int y) const;
+    virtual wxString cellValue(int x, int y) const override;
 
-    virtual double cellNumericValue(int x, int y) const;
+    virtual double cellNumericValue(int x, int y) const override;
 
-    virtual bool cellValue(int x, int y, const Yuni::String& v);
+    virtual bool cellValue(int x, int y, const Yuni::String& v) override;
 
+    /*
     virtual void resetColors(int, int, wxColour&, wxColour&) const
-    { /*Do nothing*/
+    { 
+        // Do nothing
     }
 
     virtual IRenderer::CellStyle cellStyle(int col, int row) const;
@@ -85,7 +84,9 @@ public:
     {
         return (pArea != NULL);
     }
+    */
 
+/*
 protected:
     virtual void onAreaChanged(Antares::Data::Area* area);
     virtual void onStudyClosed() override;
@@ -95,6 +96,7 @@ private:
     Antares::Data::Area* pArea;
     wxWindow* pControl;
     Toolbox::InputSelector::Area* pAreaNotifier;
+*/
 }; // class RenewableClusterSummarySingleArea
 
 } // namespace Renderer
