+++ src/ui/simulator/toolbox/components/datagrid/renderer/area/thermal.areasummary.cpp
@@ -26,8 +26,6 @@
 */
 
 #include "thermal.areasummary.h"
-#include "../../../refresh.h"
-#include "../../../../../application/study.h"
 
 using namespace Yuni;
 
@@ -40,18 +38,12 @@ namespace Datagrid
 namespace Renderer
 {
 ThermalClusterSummarySingleArea::ThermalClusterSummarySingleArea(
-  wxWindow* control,
-  Toolbox::InputSelector::Area* notifier) :
- pArea(nullptr), pControl(control), pAreaNotifier(notifier)
+    wxWindow* control,
+    Toolbox::InputSelector::Area* notifier) : CommonClusterSummarySingleArea(control, notifier)
 {
-    if (notifier)
-        notifier->onAreaChanged.connect(this, &ThermalClusterSummarySingleArea::onAreaChanged);
 }
 
-ThermalClusterSummarySingleArea::~ThermalClusterSummarySingleArea()
-{
-    destroyBoundEvents();
-}
+ThermalClusterSummarySingleArea::~ThermalClusterSummarySingleArea() {}
 
 wxString ThermalClusterSummarySingleArea::rowCaption(int rowIndx) const
 {
@@ -170,105 +162,6 @@ double ThermalClusterSummarySingleArea::cellNumericValue(int x, int y) const
     return 0.;
 }
 
-void ThermalClusterSummarySingleArea::onAreaChanged(Antares::Data::Area* area)
-{
-    if (pArea != area)
-    {
-        pArea = area;
-        RefreshAllControls(pControl);
-    }
-}
-
-IRenderer::CellStyle ThermalClusterSummarySingleArea::cellStyle(int col, int row) const
-{
-    return (col > 0 and Math::Zero(cellNumericValue(col, row)))
-             ? IRenderer::cellStyleDefaultDisabled
-             : (col == 1 || col == 2) ? IRenderer::cellStyleConstraintWeight
-                                      : IRenderer::cellStyleDefault;
-}
-
-struct NoCheck
-{
-    template<class T>
-    static bool Validate(const T&)
-    {
-        return true;
-    }
-};
-
-struct CheckMinUpDownTime
-{
-    static bool Validate(uint f)
-    {
-        return (f == 1) || (f == 24) || (f == 168);
-    }
-};
-
-struct CheckUnitCount
-{
-    static bool Validate(uint& f)
-    {
-        if (f > 100)
-            f = 100;
-        return true;
-    }
-};
-
-template<class CheckT>
-static bool UpdateUnsignedLong(uint& value, const String& str)
-{
-    uint l;
-    if (str.to(l))
-    {
-        if (value != l and CheckT::Validate(l))
-        {
-            value = l;
-            MarkTheStudyAsModified();
-            OnInspectorRefresh(nullptr);
-            return true;
-        }
-    }
-    return false;
-}
-
-template<class CheckT>
-static bool UpdateDouble(double& value, const String& str)
-{
-    double d;
-    if (str.to(d))
-    {
-        if (not Math::Equals<double>(value, d))
-        {
-            if (CheckT::Validate(d))
-            {
-                value = d;
-                OnInspectorRefresh(nullptr);
-                return true;
-            }
-        }
-    }
-    return false;
-}
-
-template<class CheckT>
-static bool UpdateBool(bool& value, const String& str)
-{
-    bool d;
-    if (str.to(d))
-    {
-        if (value != d)
-        {
-            if (CheckT::Validate(d))
-            {
-                value = d;
-                OnInspectorRefresh(nullptr);
-                return true;
-            }
-        }
-    }
-    return false;
-}
-
 bool ThermalClusterSummarySingleArea::cellValue(int x, int y, const String& v)
 {
     auto* cluster = (pArea and (uint) y < pArea->thermal.list.size())
@@ -319,18 +212,6 @@ bool ThermalClusterSummarySingleArea::cellValue(int x, int y, const String& v)
         }
     }
     return false;
-}
-
-void ThermalClusterSummarySingleArea::onStudyClosed()
-{
-    pArea = nullptr;
-    IRenderer::onStudyClosed();
-}
-
-void ThermalClusterSummarySingleArea::onStudyAreaDelete(Antares::Data::Area* area)
-{
-    if (pArea == area)
-        onAreaChanged(nullptr);
 }
 
 } // namespace Renderer
