+++ src/ui/simulator/toolbox/components/datagrid/renderer/area/thermal.areasummary.h
@@ -27,10 +27,7 @@
 #ifndef __ANTARES_TOOLBOX_COMPONENT_DATAGRID_RENDERER_AREA_THERMAL_SUMMARY_SINGLE_AREA_H__
 #define __ANTARES_TOOLBOX_COMPONENT_DATAGRID_RENDERER_AREA_THERMAL_SUMMARY_SINGLE_AREA_H__
 
-#include <antares/wx-wrapper.h>
-#include "../../renderer.h"
-#include "../../../../input/area.h"
-#include <antares/study/parts/thermal/cluster.h>
+#include "common.areasummary.h"
 
 namespace Antares
 {
@@ -40,62 +37,33 @@ namespace Datagrid
 {
 namespace Renderer
 {
-class ThermalClusterSummarySingleArea : public IRenderer
+class ThermalClusterSummarySingleArea : public CommonClusterSummarySingleArea
 {
 public:
     ThermalClusterSummarySingleArea(wxWindow* control, Toolbox::InputSelector::Area* notifier);
 
     virtual ~ThermalClusterSummarySingleArea();
 
-    virtual int width() const
+    virtual int width() const override
     {
         return 14;
     }
-    virtual int height() const
+    virtual int height() const override
     {
         return (pArea) ? pArea->thermal.list.size() : 0;
     }
 
-    virtual wxString columnCaption(int colIndx) const;
+    virtual wxString columnCaption(int colIndx) const override;
 
-    virtual wxString rowCaption(int rowIndx) const;
+    virtual wxString rowCaption(int rowIndx) const override;
 
-    virtual wxString cellValue(int x, int y) const;
+    virtual wxString cellValue(int x, int y) const override;
 
-    virtual double cellNumericValue(int x, int y) const;
+    virtual double cellNumericValue(int x, int y) const override;
 
-    virtual bool cellValue(int x, int y, const Yuni::String& v);
+    virtual bool cellValue(int x, int y, const Yuni::String& v) override;
 
-    virtual void resetColors(int, int, wxColour&, wxColour&) const
-    { /*Do nothing*/
-    }
-
-    virtual IRenderer::CellStyle cellStyle(int col, int row) const;
-
-    virtual uint maxWidthResize() const
-    {
-        return 0;
-    }
-    virtual uint maxHeightResize() const
-    {
-        return 0;
-    }
-
-    virtual bool valid() const
-    {
-        return (pArea != NULL);
-    }
-
-protected:
-    virtual void onAreaChanged(Antares::Data::Area* area);
-    virtual void onStudyClosed() override;
-    void onStudyAreaDelete(Antares::Data::Area* area);
-
-private:
-    Antares::Data::Area* pArea;
-    wxWindow* pControl;
-    Toolbox::InputSelector::Area* pAreaNotifier;
-}; // class ThermalClusterPrepro
+}; // class ThermalClusterSummarySingleArea
 
 } // namespace Renderer
 } // namespace Datagrid
